backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic[email]==2.6.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==7.15.0
google-genai==1.41.0
pytest==8.0.0
httpx==0.26.0
